+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.12](https://github.com/rfsbraz/deleterr/compare/v0.2.11...v0.2.12) (2026-03-05)
+
+
+### Bug Fixes
+
+* **tautulli:** match TV show watch data by rating key ([#248](https://github.com/rfsbraz/deleterr/issues/248)) ([214262f](https://github.com/rfsbraz/deleterr/commit/214262fb40641b7d7c0d2f9f5aba7ab1f0a856ba))
+
+## [0.2.11](https://github.com/rfsbraz/deleterr/compare/v0.2.10...v0.2.11) (2026-03-04)
+
+
+### Features
+
+* **overseerr:** protect content unwatched by requester ([#244](https://github.com/rfsbraz/deleterr/issues/244)) ([685d997](https://github.com/rfsbraz/deleterr/commit/685d99758a3552c79f653c1efc7b5c3c0f56ca0b))
+
 ## [0.2.10](https://github.com/rfsbraz/deleterr/compare/v0.2.9...v0.2.10) (2026-02-23)